Add NaturalModuleIdsPlugin
