+++ lib/Compilation.js
@@ -129,21 +129,6 @@ const byIdOrIdentifier = (a, b) => {
 };
 
 /**
- * @param {Module} a first module to sort by
- * @param {Module} b second module to sort by
- * @returns {-1|0|1} sort value
- */
-const byIndexOrIdentifier = (a, b) => {
-	if (a.index < b.index) return -1;
-	if (a.index > b.index) return 1;
-	const identA = a.identifier();
-	const identB = b.identifier();
-	if (identA < identB) return -1;
-	if (identA > identB) return 1;
-	return 0;
-};
-
-/**
  * @param {Compilation} a first compilation to sort by
  * @param {Compilation} b second compilation to sort by
  * @returns {-1|0|1} sort value
@@ -1153,7 +1138,6 @@ class Compilation {
 			this.assignDepth(module);
 		}
 		this.processDependenciesBlocksForChunkGroups(this.chunkGroups.slice());
-		this.sortModules(this.modules);
 		this.hooks.afterChunks.call(this.chunks);
 
 		this.hooks.optimize.call();
@@ -1262,17 +1246,6 @@ class Compilation {
 				});
 			});
 		});
-	}
-
-	/**
-	 * @param {Module[]} modules the modules array on compilation to perform the sort for
-	 * @returns {void}
-	 */
-	sortModules(modules) {
-		// TODO webpack 5: this should only be enabled when `moduleIds: "natural"`
-		// TODO move it into a plugin (NaturalModuleIdsPlugin) and use this in WebpackOptionsApply
-		// TODO remove this method
-		modules.sort(byIndexOrIdentifier);
 	}
 
 	/**
