+++ lib/WebpackOptionsApply.js
@@ -58,6 +58,7 @@ const RuntimeChunkPlugin = require("./optimize/RuntimeChunkPlugin");
 const NoEmitOnErrorsPlugin = require("./NoEmitOnErrorsPlugin");
 const NamedModulesPlugin = require("./NamedModulesPlugin");
 const NamedChunksPlugin = require("./NamedChunksPlugin");
+const NaturalModuleIdsPlugin = require("./NaturalModuleIdsPlugin");
 const HashedModuleIdsPlugin = require("./HashedModuleIdsPlugin");
 const DefinePlugin = require("./DefinePlugin");
 const SizeLimitsPlugin = require("./performance/SizeLimitsPlugin");
@@ -359,7 +360,7 @@ class WebpackOptionsApply extends OptionsApply {
 		if (moduleIds) {
 			switch (moduleIds) {
 				case "natural":
-					// TODO webpack 5: see hint in Compilation.sortModules
+					new NaturalModuleIdsPlugin().apply(compiler);
 					break;
 				case "named":
 					new NamedModulesPlugin().apply(compiler);
